+++ crates/bitwarden-crypto/src/enc_string/symmetric.rs
@@ -362,7 +362,7 @@ impl EncString {
         key: &XChaCha20Poly1305Key,
         content_format: ContentFormat,
     ) -> Result<EncString> {
-        let data = crate::cose::encrypt_xchacha20_poly1305(data_dec, key, content_format)?;
+        let data = crate::cose::symmetric::encrypt_xchacha20_poly1305(data_dec, key, content_format)?;
         Ok(EncString::Cose_Encrypt0_B64 {
             data: data.to_vec(),
         })
@@ -419,7 +419,7 @@ impl KeyDecryptable<SymmetricCryptoKey, Vec<u8>> for EncString {
                 EncString::Cose_Encrypt0_B64 { data },
                 SymmetricCryptoKey::XChaCha20Poly1305Key(key),
             ) => {
-                let (decrypted_message, _) = crate::cose::decrypt_xchacha20_poly1305(
+                let (decrypted_message, _) = crate::cose::symmetric::decrypt_xchacha20_poly1305(
                     &CoseEncrypt0Bytes::from(data.as_slice()),
                     key,
                 )?;
